feat : 총 상금을 계산하는 기능

+++ src/main/java/lotto/domain/Rank.java
@@ -9,7 +9,7 @@ public enum Rank {
     FIFTH_PRIZE(3, 5_000, "3개 일치 (5,000원)");
 
     private final int matchingCount;
-    private final int prize;
+    private final long prize;
     private final String description;
 
     Rank(int matchingCount, int prize, String description) {
@@ -18,7 +18,7 @@ public enum Rank {
         this.description = description;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
